+++ dmr/openapi/dump.py
@@ -3,13 +3,12 @@ from typing import TYPE_CHECKING
 from dmr.internal.json import json_dump as _json_dump
 
 if TYPE_CHECKING:
-    from dmr.openapi.objects.openapi import ConvertedSchema
-    from dmr.openapi.views.base import DumpedSchema
+    from dmr.openapi.mappers.schema_normalization import DumpedSchema
 
 
-def json_dump(schema: 'ConvertedSchema') -> 'DumpedSchema':
+def json_dump(schema: 'DumpedSchema') -> str:
     """
-    Serialize `ConvertedSchema` to a decoded JSON string.
+    Serialize `DumpedSchema` to a decoded JSON string.
 
     Args:
         schema: Converted OpenAPI schema to be serialized.