+++ dmr/openapi/objects/reference.py
@@ -1,4 +1,7 @@
 from dataclasses import dataclass
+from typing import Annotated
+
+from dmr.internal.dataclass_aliases import Field
 
 
 @dataclass(kw_only=True, slots=True)
@@ -10,6 +13,6 @@ class Reference:
     the location of the value being referenced.
     """
 
-    ref: str
+    ref: Annotated[str, Field(alias='$ref')]
     summary: str | None = None
     description: str | None = None