+++ dmr/plugins/msgspec/serializer.py
@@ -124,13 +124,16 @@ class MsgspecSerializer(BaseSerializer):
                 For example, it is fine for a request validation
                 to be less strict in some cases and allow type coercition.
                 But, response types need to be strongly validated.
-            extra_namespace: Is not used.
+            extra_namespace: Not used currently.
 
         Returns:
             Structured and validated data.
 
         Raises:
             msgspec.ValidationError: When parsing can't be done.
+
+        .. versionchanged:: 0.13.0
+            Added *extra_namespace* parameter.
 
         """
         return msgspec.convert(