+++ CHANGELOG.md
@@ -18,11 +18,32 @@ Later on we will make the API more stable and decrease the amount
 of requirements for an API to count as public.
 
 
-## WIP
+## 0.13.0 WIP
+
+### Breaking changes
+
+- `Schema.then` is renamed to be `Schema.schema_then`
+  to be consistent with other similar names, #1221
+- `dmr.openapi.objects.openapi.convert` function is renamed and moved
+  to `dmr.openapi.mappers.schema_normalization.dump_schema`, #1221
+- `dmr.openapi.objects.openapi.normalize_key`
+  and `dmr.openapi.objects.openapi.normalize_value` functions are removed, #1221
+- `dmr.openapi.objects.openapi.ConvertedSchema` is renamed and moved
+  to `dmr.openapi.mappers.schema_normalization.DumpedSchema`, #1221
+- `dmr.openapi.views.base.DumpedSchema` is removed,
+  it was just a `str` type alias, #1221
+- `dmr.openapi.objects.OpenAPI` is moved
+  to `dmr.openapi.openapi.OpenAPI`, #1222
+- `rebuild_namespace` parameter in `PydanticSerializer.from_python`
+  was renamed to `extra_namespace`, #1222
 
 ### Features
 
-- Django 6.1 official support
+- Django 6.1 official support, #1214
+- Added `extra_namespace` parameter to `BaseSerializer.from_python`
+  and all its existing subclasses, #1222
+- Added an ability to load external OpenAPI schemas
+  into our typed dataclasses, #1222
 
 
 ## 0.12.1 (2026-07-31)